+++ CHANGELOG.md
@@ -1,4 +1,14 @@
 # Changelog
 
+## 1.0.1
+* Added application icon.
+* Added a Configure window, which will allow the user to customize script types that can be ran:
+  * Name of type.
+  * Extension(s) for type.
+  * Executable for type (e.g. python.exe path).
+  * Customizable command format.
+  * Toggle if the terminal should be visible.
+* Changed save data, which is now stored in a json file, in the local application data folder.
+
 ## 1.0.0
 * Initial release.